requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
insta-scrape==2.1.2
moviepy==1.0.3
numpy==1.26.4
openai==1.14.2
pillow==10.2.0
playwright==1.42.0
proglog==0.1.10
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.0.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
